// src/app/models/useCase.ts
export class UseCase {
    cityId: string;
    zoneId: string;
    zoneSymbol: string;
    zoneDescription: string;
    applicationURL: string;
    procedureURL: string;
    permitName:string;
    permitDescription: string;
    zoneLandUseId: string;
    zoneLandUseDescription: string;
    generalStandardURL: string;
    gardenStandardURL: string;
    additionalStandardURL: string;
    frontageAndFacadesStandardURL: string;
  }